Extract cdk-dependent schematics
So we can fail gracefully if material is missing
